+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.2] - 2024-07-11
 
+### Changed
+
+- dropped 'verbose' argument from list_boards()
 
 ## [0.1.1] - 2024-07-11